backend/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.2
gensim==4.3.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.3
numpy==2.2.3
pandas==2.2.3
PyPDF2==3.0.1
scikit-learn==1.6.1
spacy==3.8.4
python-docx==1.1.2
Werkzeug==3.1.5
